+++ app/controllers/residents_controller.rb
@@ -71,8 +71,8 @@ class ResidentsController < ApplicationController
         :middle_name,
         :first_name,
         :email_address,
+        :phone,
         :is_minor,
-        :age_of_minor,
         residencies_attributes: [
           :id,
           :resident_id,