+++ app/decorators/lot_decorator.rb
@@ -5,6 +5,10 @@ class LotDecorator < Draper::Decorator
     object.lot_number || "⁇"
   end
 
+  def paid_on
+    h.date_tag(object.paid_on)
+  end
+
   def residents_summary(type: :icons)
     h.render 'residents/resident_icon_list', residents: object.residents.decorate
   end