+++ app/decorators/resident_decorator.rb
@@ -5,7 +5,11 @@ class ResidentDecorator < Draper::Decorator
     'user'
   end
 
+  def phone
+    h.number_to_phone(object.phone, area_code: true)
+  end
+
   def property_summary(type: :icons)
-    h.render 'properties/property_icon_list', properties: object.properties.decorate
+    h.render 'residencies/property_icon_list', residencies: resident.residencies.decorate
   end
 end