+++ db/migrate/20220521051327_add_phone_to_residents.rb
@@ -1,0 +1,5 @@
+class AddPhoneToResidents < ActiveRecord::Migration[7.0]
+  def change
+    add_column :residents, :phone, :string
+  end
+end