+++ db/migrate/20220521172110_remove_age_of_minor.rb
@@ -1,0 +1,5 @@
+class RemoveAgeOfMinor < ActiveRecord::Migration[7.0]
+  def change
+    remove_column :residents, :age_of_minor, :text
+  end
+end