+++ db/migrate/20220522223332_convert_residency_resident_status_to_string.rb
@@ -1,0 +1,9 @@
+class ConvertResidencyResidentStatusToString < ActiveRecord::Migration[7.0]
+  def up
+    change_column :residencies, :resident_status, :string
+  end
+
+  def down
+    change_column :residencies, :resident_status, :integer
+  end
+end